Orrery - Moons
I messed with a lot of trig for a while because I thought I needed to make the planets orbit the sun on a sphere, instead of the 2d circe orbit path. Moons added but rotation/radius need to be completed on all of them. I want the game to save the transform data + game time in an array, with a set amount of forward and backwards time it stores.